Restructure FullTransactionInfo module classes
