+++ app/src/main/AndroidManifest.xml
@@ -22,9 +22,9 @@
             android:launchMode="singleTask"
             android:screenOrientation="portrait">
             <intent-filter>
-                <action android:name="android.intent.action.MAIN"/>
+                <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER"/>
+                <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
         <activity
@@ -45,7 +45,7 @@
         <activity
             android:name=".modules.send.SendActivity"
             android:screenOrientation="portrait"
-            android:windowSoftInputMode="stateVisible"/>
+            android:windowSoftInputMode="stateVisible" />
         <activity
             android:name=".modules.restore.words.RestoreWordsActivity"
             android:screenOrientation="portrait" />
@@ -53,7 +53,7 @@
             android:name=".modules.qrscanner.QRScannerActivity"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".modules.fulltransactioninfo.FullTransactionInfoActivity"
+            android:name=".modules.fulltransactioninfo.views.FullTransactionInfoActivity"
             android:screenOrientation="portrait" />
         <activity
             android:name=".modules.fulltransactioninfo.dataprovider.DataProviderSettingsActivity"
