+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/fulltransactioninfo/FullTransactionInfoFactory.kt
@@ -4,6 +4,9 @@ import io.horizontalsystems.bankwallet.core.INetworkManager
 import io.horizontalsystems.bankwallet.core.ITransactionDataProviderManager
 import io.horizontalsystems.bankwallet.entities.Coin
 import io.horizontalsystems.bankwallet.entities.CoinType
+import io.horizontalsystems.bankwallet.modules.fulltransactioninfo.views.FullTransactionBinanceAdapter
+import io.horizontalsystems.bankwallet.modules.fulltransactioninfo.views.FullTransactionBitcoinAdapter
+import io.horizontalsystems.bankwallet.modules.fulltransactioninfo.views.FullTransactionEthereumAdapter
 
 class FullTransactionInfoFactory(private val networkManager: INetworkManager, private val dataProviderManager: ITransactionDataProviderManager)
     : FullTransactionInfoModule.ProviderFactory {
