+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/fulltransactioninfo/providers/BlockdozerBitcoinCashProvider.kt
@@ -2,10 +2,9 @@ package io.horizontalsystems.bankwallet.modules.fulltransactioninfo.providers
 
 import com.google.gson.Gson
 import com.google.gson.JsonObject
-import io.horizontalsystems.bankwallet.modules.fulltransactioninfo.BitcoinResponse
 import io.horizontalsystems.bankwallet.modules.fulltransactioninfo.FullTransactionInfoModule
 
-class BlockdozerBitcoinCashProvider(private val testMode: Boolean): FullTransactionInfoModule.BitcoinForksProvider {
+class BlockdozerBitcoinCashProvider(testMode: Boolean) : FullTransactionInfoModule.BitcoinForksProvider {
     override val name = "Blockdozer.com"
 
     private val baseUrl = "https://${if (testMode) "tbch." else ""}blockdozer.com"
