+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/fulltransactioninfo/providers/InsightDashProvider.kt
@@ -2,7 +2,6 @@ package io.horizontalsystems.bankwallet.modules.fulltransactioninfo.providers
 
 import com.google.gson.Gson
 import com.google.gson.JsonObject
-import io.horizontalsystems.bankwallet.modules.fulltransactioninfo.BitcoinResponse
 import io.horizontalsystems.bankwallet.modules.fulltransactioninfo.FullTransactionInfoModule
 
 class InsightDashProvider : FullTransactionInfoModule.BitcoinForksProvider {
