+++ app/src/main/res/layout/fragment_full_transaction_info_error.xml
@@ -1,9 +1,9 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                                                   xmlns:app="http://schemas.android.com/apk/res-auto"
-                                                   xmlns:tools="http://schemas.android.com/tools"
-                                                   android:layout_width="match_parent"
-                                                   android:layout_height="match_parent"
-                                                   tools:context="io.horizontalsystems.bankwallet.modules.fulltransactioninfo.FullTransactionInfoErrorFragment">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="io.horizontalsystems.bankwallet.modules.fulltransactioninfo.views.FullTransactionInfoErrorFragment">
 
     <ImageView
         android:id="@+id/imageView"
@@ -25,8 +25,8 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="24dp"
         android:layout_marginEnd="8dp"
-        android:textColor="?TextColorBarsToDark"
         android:text="@string/FullInfo_Source"
+        android:textColor="?TextColorBarsToDark"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.512"
         app:layout_constraintStart_toStartOf="parent"
@@ -66,8 +66,8 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="24dp"
-        android:textColor="?TextColorGreyLightToDark"
         android:text="@string/FullInfo_Error_ChangeSource"
+        android:textColor="?TextColorGreyLightToDark"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/btnRetry" />
