+++ app/src/main/res/layout/view_holder_full_transaction_item.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<io.horizontalsystems.bankwallet.modules.fulltransactioninfo.FullTransactionInfoItemView
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<io.horizontalsystems.bankwallet.modules.fulltransactioninfo.views.FullTransactionInfoItemView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/sectionItem"
     android:layout_width="match_parent"
